Add buildLatencyMetrics method to HistogramMetrics
