+++ connector/spanmetricsconnector/connector_test.go
@@ -22,7 +22,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/lightstep/go-expohisto/structure"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
@@ -817,7 +816,7 @@ func newConnectorImp(
 
 		startTimestamp: pcommon.NewTimestampFromTime(time.Now()),
 		histograms:     histogramMetrics,
-		sums:           make(map[string]*Sum),
+		sums:           SumMetrics{Metrics: make(map[string]*Sum)},
 		dimensions: []dimension{
 			// Set nil defaults to force a lookup for the attribute in the span.
 			{stringAttrName, nil},
@@ -982,98 +981,5 @@ func TestBuildMetricName(t *testing.T) {
 	for _, test := range tests {
 		actual := buildMetricName(test.namespace, test.metricName)
 		assert.Equal(t, test.expected, actual)
-	}
-}
-
-func TestConnector_ExpoHistToExponentialDataPoint(t *testing.T) {
-	tests := []struct {
-		name  string
-		input *structure.Histogram[float64]
-		want  pmetric.ExponentialHistogramDataPoint
-	}{
-		{
-			name:  "max bucket size - 4",
-			input: structure.NewFloat64(structure.NewConfig(structure.WithMaxSize(4)), 2, 4),
-			want: func() pmetric.ExponentialHistogramDataPoint {
-				dp := pmetric.NewExponentialHistogramDataPoint()
-				dp.SetCount(2)
-				dp.SetSum(6)
-				dp.SetMin(2)
-				dp.SetMax(4)
-				dp.SetZeroCount(0)
-				dp.SetScale(1)
-				dp.Positive().SetOffset(1)
-				dp.Positive().BucketCounts().FromRaw([]uint64{
-					1, 0, 1,
-				})
-				return dp
-			}(),
-		},
-		{
-			name:  "max bucket size - default",
-			input: structure.NewFloat64(structure.NewConfig(), 2, 4),
-			want: func() pmetric.ExponentialHistogramDataPoint {
-				dp := pmetric.NewExponentialHistogramDataPoint()
-				dp.SetCount(2)
-				dp.SetSum(6)
-				dp.SetMin(2)
-				dp.SetMax(4)
-				dp.SetZeroCount(0)
-				dp.SetScale(7)
-				dp.Positive().SetOffset(127)
-				buckets := make([]uint64, 129)
-				buckets[0] = 1
-				buckets[128] = 1
-				dp.Positive().BucketCounts().FromRaw(buckets)
-				return dp
-			}(),
-		},
-		{
-			name:  "max bucket size - 4, negative observations",
-			input: structure.NewFloat64(structure.NewConfig(structure.WithMaxSize(4)), -2, -4),
-			want: func() pmetric.ExponentialHistogramDataPoint {
-				dp := pmetric.NewExponentialHistogramDataPoint()
-				dp.SetCount(2)
-				dp.SetSum(-6)
-				dp.SetMin(-4)
-				dp.SetMax(-2)
-				dp.SetZeroCount(0)
-				dp.SetScale(1)
-				dp.Negative().SetOffset(1)
-				dp.Negative().BucketCounts().FromRaw([]uint64{
-					1, 0, 1,
-				})
-				return dp
-			}(),
-		},
-		{
-			name:  "max bucket size - 4, negative and positive observations",
-			input: structure.NewFloat64(structure.NewConfig(structure.WithMaxSize(4)), 2, 4, -2, -4),
-			want: func() pmetric.ExponentialHistogramDataPoint {
-				dp := pmetric.NewExponentialHistogramDataPoint()
-				dp.SetCount(4)
-				dp.SetSum(0)
-				dp.SetMin(-4)
-				dp.SetMax(4)
-				dp.SetZeroCount(0)
-				dp.SetScale(1)
-				dp.Positive().SetOffset(1)
-				dp.Positive().BucketCounts().FromRaw([]uint64{
-					1, 0, 1,
-				})
-				dp.Negative().SetOffset(1)
-				dp.Negative().BucketCounts().FromRaw([]uint64{
-					1, 0, 1,
-				})
-				return dp
-			}(),
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := pmetric.NewExponentialHistogramDataPoint()
-			expoHistToExponentialDataPoint(tt.input, got)
-			assert.Equal(t, tt.want, got)
-		})
 	}
 }
